requirements.txt
psycopg2-binary
loguru
kafka-python
requests